use last known location if it is recent enough
